pyproject.toml
# Project basic information configuration, required.
[project]
name = "datus-agent"
version = "0.2.4"
description = "AI-powered SQL Agent for data engineering (Compiled Version)"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
license-files = ["LICENSE*"]
authors = [
    { name = "Datus Team", email = "harrison.zhao@datus.ai" }
]
maintainers = [
    { name = "Datus Team", email = "harrison.zhao@datus.ai" }
]
keywords = ["sql", "ai", "agent", "database", "nlp", "natural-language"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-dotenv==1.0.0",
    "pandas==2.1.4",
    "sqlalchemy==2.0.23",
    "sqlglot>=26.12.0",
    "pyyaml==6.0.1",
    "structlog>=23.1.0",
    "openai>=2.8.0",
    "httpx[socks]==0.27.2",
    "tantivy>=0.22.2",
    "aiohttp>=3.11.16",
    "xlsxwriter>=3.2.2",
    "openai-agents[litellm]==0.7.0",
    "litellm>=1.67.4.post1,<2",
    "pydantic>=2.11.7,<3.0",
    "lancedb==0.18.0",
    # This library is required by lancedb
    # this verison is required by pandas 2.1.4
    "pyarrow<19.0.0",
    "rich==14.0.0",
    "prompt_toolkit>=3.0.51",
    "pygments>=2.18.0",
    "textual[syntax]==5.1.1",
    "anthropic==0.51.0",
    "duckdb-engine>=0.17.0",
    "duckdb==1.3.0",
    # MCP (Model Context Protocol) support
    "mcp>=1.11.0",
    "anyio>=4.9.0",
    "json-repair>=0.47.6",
    "fastapi>=0.104.0,<1.0",
    "uvicorn>=0.24.0",
    "google-generativeai>=0.8.0",
    "pyperclip==1.9.0",
    "streamlit>=1.38.0",
    "wcmatch>=10.0",
    "fastembed==0.4.2",
    # Optional: install torch manually to enable reranker models.
    "uv>=0.9.9",
    "plotly>=6.5.0",
    "charset-normalizer>=3.4.2",
    # Platform documentation tools
    "PyGithub>=2.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "markdown-it-py>=3.0.0",
]

[project.optional-dependencies]
# Database connectors have been moved to separate adapter packages
# See https://github.com/Datus-ai/Datus-adapters for available adapters

[project.urls]
Homepage = "https://datus.ai/"
Documentation = "https://github.com/datus-ai/datus-agent#readme"
Repository = "https://github.com/datus-ai/datus-agent"
"Bug Tracker" = "https://github.com/datus-ai/datus-agent/issues"

[project.scripts]
datus-agent = "datus.main:main"
datus-cli = "datus.cli.main:main"
datus = "datus.cli.main:main"
datus-mcp = "datus.mcp_server:main"

[[tool.uv.index]]
url = "http://mirrors.aliyun.com/pypi/simple"
default = true

[build-system]
requires = ["setuptools>=80.9.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["datus*"]
exclude = ["benchmark*", "docs*"]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json", "*.md", "*.txt", "*.pyc", "*.j2"]
"datus" = ["*.yml", "*.yaml", "*.json", "*.md", "*.txt"]
"datus.prompts" = ["*.txt", "*.md", "*.j2"]
"datus.prompts.prompt_templates" = ["*.j2"]

[tool.setuptools.package-dir]
"mcp.mcp-metricflow-server" = "mcp/mcp-metricflow-server"

[tool.setuptools.data-files]
"datus/conf" = ["conf/*.yml.*"]
"datus/sample_data" = ["sample_data/duckdb-demo.duckdb"]
"datus/sample_data/california_schools" = [
    "sample_data/california_schools/california_schools.csv",
    "sample_data/california_schools/success_story.csv",
    "sample_data/california_schools/california_schools.sqlite"
]
"datus/sample_data/california_schools/reference_sql" = [
    "sample_data/california_schools/reference_sql/*.sql",
]
"datus/sample_data/superset" = [
    "sample_data/superset/*.*",
]

[tool.uv.pip]
index-url = "http://mirrors.aliyun.com/pypi/simple"

# uv tool run black
[tool.black]
line-length = 120
target-version = ["py312"]

[tool.flake8]
max-line-length = 120
ignore = "E203, E266, E501, W503"
max-complexity = 10
select = "B,C,E,F,W,T4,B9"
exclude = "mcp/"

[tool.pytest]
pythonpath = "."
testpaths = ["tests"]

[dependency-groups]
# uv sync --dev
dev = [
    "pytest==8.0.2",
    "pytest-asyncio>=0.23.8",
    "black==23.12.1",
    "mypy==1.8.0",
    "tqdm>=4.27.0",
    "pytest-xdist>=3.7.0",
    "pre-commit>=4.2.0",
    "setuptools>=80.9.0",
    "pip>=25.1.1",
    "build>=1.2.0",
    "twine>=6.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "langsmith>=0.0.77",
    "langsmith-fetch>=0.1.0",
]

[tool.coverage.run]
source = ["datus"]
omit = [
    "datus/mcp_server.py",
    "datus/cli/*",
    "datus/api/*",
    "datus/prompts/prompt_templates/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
    "except ImportError",
]
show_missing = true
precision = 2
